backend/requirements.txt
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
pytest-django==4.4.0
django-environ==0.8.1
psycopg2-binary==2.9.1